+++ CHANGELOG.md
@@ -13,6 +13,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2092 Migrate trends target range lines to dumb
 - YLP-2094 Update some dependencies with major changes
 - YLP-2095 Migrate FocusedCBGSliceSegment from viz to dumb
+- YLP-2097 Migrate RangeSelect from viz to dumb
 
 ## 3.1.0 - 2023-01-31
 ### Fixed